Suppression des volumes dans `docker-compose.yml` pour éviter la persistance des données inutiles. Mise à jour des répertoires surveillés dans `main.py` pour exclure `feedback_data` et `exports`, optimisant ainsi le rechargement. Refactorisation de `lifespan.py` pour utiliser `initialize_database` depuis `db.session` au lieu de `ui.db_init`, simplifiant l'initialisation de la base de données. Suppression du fichier `db_init.py` obsolète.
