+++ main.py
@@ -54,13 +54,11 @@ try:
                 host="0.0.0.0",
                 port=settings.agent.AGENT_PORT,
                 reload=True,
-                reload_dirs=["src", "static"],
+                reload_dirs=["src"],
                 reload_excludes=[
                     "*.pyc",
                     "__pycache__",
                     "*.log",
-                    "feedback_data",
-                    "exports",
                 ],
                 log_level="info",
                 access_log=True,
